Переделана структура каталогов для заданий по расписанию. В Startup.cs их создание вынесено в отдельный метод
