+++ MetricsManager/MetricsAgent/ScheduledWorks/Jobs/DotNetMetricJob.cs
@@ -5,31 +5,31 @@ using System;
 using System.Diagnostics;
 using System.Threading.Tasks;
 
-namespace MetricsAgent.Jobs
+namespace MetricsAgent.ScheduledWorks
 {
 	/// <summary>
-	/// Задача сбора Cpu метрик
+	/// Задача сбора DotNet метрик
 	/// </summary>
-	public class CpuMetricJob : IJob
+	public class DotNetMetricJob : IJob
 	{
 		// Инжектируем DI провайдер
 		private readonly IServiceProvider _provider;
-		private ICpuMetricsRepository _repository;
+		private IDotNetMetricsRepository _repository;
 
 		/// <summary>Имя категории счетчика</summary>
-		private readonly string categoryName = "Processor";
+		private readonly string categoryName = ".NET CLR Memory";
 		/// <summary>Имя счетчика</summary>
-		private readonly string counterName = "% Processor Time";
+		private readonly string counterName = "# Bytes in all Heaps";
 		/// <summary>Счетчик</summary>
 		private PerformanceCounter _counter;
 
 
-		public CpuMetricJob(IServiceProvider provider)
+		public DotNetMetricJob(IServiceProvider provider)
 		{
 			_provider = provider;
-			_repository = _provider.GetService<ICpuMetricsRepository>();
+			_repository = _provider.GetService<IDotNetMetricsRepository>();
 
-			_counter = new PerformanceCounter(categoryName, counterName, "_Total");
+			_counter = new PerformanceCounter(categoryName, counterName, "_Global_");
 
 		}
 
@@ -42,7 +42,7 @@ namespace MetricsAgent.Jobs
 			var time = TimeSpan.FromSeconds(DateTimeOffset.UtcNow.ToUnixTimeSeconds());
 
 			// Запись метрики в репозиторий
-			_repository.Create(new CpuMetric { Time = time, Value = value });
+			_repository.Create(new DotNetMetric { Time = time, Value = value });
 
 			return Task.CompletedTask;
 		}
