+++ MetricsManager/MetricsAgent/ScheduledWorks/Tools/QuartzHostedService.cs
