+++ MetricsManager/MetricsAgent/Startup.cs
@@ -12,7 +12,7 @@ using AutoMapper;
 using FluentMigrator.Runner;
 using Quartz;
 using Quartz.Spi;
-using MetricsAgent.Jobs;
+using MetricsAgent.ScheduledWorks;
 using Quartz.Impl;
 using MetricsAgent.SQLsettings;
 
@@ -21,9 +21,9 @@ namespace MetricsAgent
 	public class Startup
 	{
 		/// <summary>
-		/// Строка для подключения к базе данных
+		/// Периодичность запуска задач по сбору метрик
 		/// </summary>
-		private const string ConnectionString = @"Data Source=metrics.db; Version=3;Pooling=True;Max Pool Size=100;";
+		private const string CronExpression = "0/5 * * * * ?";
 
 		public Startup(IConfiguration configuration)
 		{
@@ -35,51 +35,70 @@ namespace MetricsAgent
 		// This method gets called by the runtime. Use this method to add services to the container.
 		public void ConfigureServices(IServiceCollection services)
 		{
+			// Контроллеры
 			services.AddControllers();
 
+			// Репозитории
 			services.AddSingleton<ICpuMetricsRepository, CpuMetricsRepository>();
 			services.AddSingleton<IDotNetMetricsRepository, DotNetMetricsRepository>();
 			services.AddSingleton<IHddMetricsRepository, HddMetricsRepository>();
 			services.AddSingleton<INetworkMetricsRepository, NetworkMetricsRepository>();
 			services.AddSingleton<IRamMetricsRepository, RamMetricsRepository>();
 
+			// Маппер
 			var mapperConfiguration = new MapperConfiguration(mp => mp.AddProfile(new MapperProfile()));
 			var mapper = mapperConfiguration.CreateMapper();
 			services.AddSingleton(mapper);
 
+			// Настройки для базы данных
 			services.AddSingleton<IMySqlSettings, MySqlSettings>();
 
+			// Мигратор
 			services.AddFluentMigratorCore()
 				.ConfigureRunner(rb => rb.AddSQLite() // добавляем поддержку SQLite 
-					.WithGlobalConnectionString(ConnectionString)// устанавливаем строку подключения
+					.WithGlobalConnectionString(new MySqlSettings().ConnectionString)// устанавливаем строку подключения
 					.ScanIn(typeof(Startup).Assembly).For.Migrations())// подсказываем где искать классы с миграциями
 				.AddLogging(lb => lb.AddFluentMigratorConsole());
 
+			// Настройка сбора метрик по расписанию
+			JobsSheduleSettings(services);
+		}
+
+		/// <summary>
+		/// Настройка сбора метрик по расписанию
+		/// </summary>
+		/// <param name="services"></param>
+		private void JobsSheduleSettings(IServiceCollection services)
+		{
+			// Планировщики заданий
 			services.AddSingleton<IJobFactory, SingletonJobFactory>();
 			services.AddSingleton<ISchedulerFactory, StdSchedulerFactory>();
 
+			// Задачи для разных метрик
 			services.AddSingleton<CpuMetricJob>();
 			services.AddSingleton<DotNetMetricJob>();
 			services.AddSingleton<HddMetricJob>();
 			services.AddSingleton<NetworkMetricJob>();
 			services.AddSingleton<RamMetricJob>();
 
+			// Периодичность запуска задач
 			services.AddSingleton(new JobSchedule(
 				jobType: typeof(CpuMetricJob),
-				cronExpression: "0/5 * * * * ?"));
+				cronExpression: CronExpression));
 			services.AddSingleton(new JobSchedule(
 				jobType: typeof(DotNetMetricJob),
-				cronExpression: "0/5 * * * * ?"));
+				cronExpression: CronExpression));
 			services.AddSingleton(new JobSchedule(
 				jobType: typeof(HddMetricJob),
-				cronExpression: "0/5 * * * * ?"));
+				cronExpression: CronExpression));
 			services.AddSingleton(new JobSchedule(
 				jobType: typeof(NetworkMetricJob),
-				cronExpression: "0/5 * * * * ?"));
+				cronExpression: CronExpression));
 			services.AddSingleton(new JobSchedule(
 				jobType: typeof(RamMetricJob),
-				cronExpression: "0/5 * * * * ?"));
+				cronExpression: CronExpression));
 
+			// Сервис для запуска задач с помощью Quarz
 			services.AddHostedService<QuartzHostedService>();
 		}
